--- app/src/main/java/space/taran/arknavigator/mvp/presenter/adapter/IResourcesGridPresenter.kt
@@ -1,10 +1,0 @@
-package space.taran.arknavigator.mvp.presenter.adapter
-
-import space.taran.arknavigator.mvp.model.dao.ResourceId
-import space.taran.arknavigator.mvp.view.item.FileItemView
-
-interface IResourcesGridPresenter {
-    fun getCount(): Int
-    fun bindView(view: FileItemView)
-    fun onItemClick(pos: Int)
-}
